pyproject.toml
[tool.poetry]
name = "post_backend"
version = "0.1.0"
description = ""
authors = ["yakubov <yakubov9791999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "3.13.1"
djoser = "2.1.0"
djangorestframework-simplejwt = "4.8.0"
django-cors-headers = "^3.11.0"
python-dotenv = "^0.19.2"
psycopg2-binary = "^2.9.3"
Pillow = "^9.0.1"
django-filter = "^21.1"
asgiref = "^3.5.0"
attrs = "^21.4.0"
CacheControl = "^0.12.11"
cachy = "^0.3.0"
certifi = "^2021.10.8"
cffi = "^1.15.0"
charset-normalizer = "^2.0.12"
cleo = "^0.8.1"
clikit = "^0.6.2"
coreapi = "^2.3.3"
coreschema = "^0.0.4"
coverage = "^6.3.2"
crashtest = "^0.3.1"
cryptography = "^36.0.2"
defusedxml = "^0.7.1"
distlib = "^0.3.4"
Django = "^4.0.4"
django-templated-mail = "^1.1.1"
factory-boy = "^3.2.1"
Faker = "^13.4.0"
filelock = "^3.6.0"
html5lib = "^1.1"
idna = "^3.3"
importlib-metadata = "^4.11.3"
inflection = "^0.5.1"
iniconfig = "^1.1.1"
itypes = "^1.2.0"
jeepney = "^0.8.0"
Jinja2 = "^3.1.1"
keyring = "^23.5.0"
lockfile = "^0.12.2"
MarkupSafe = "^2.1.1"
msgpack = "^1.0.3"
oauthlib = "^3.2.0"
packaging = "^21.3"
pastel = "^0.2.1"
pexpect = "^4.8.0"
pkginfo = "^1.8.2"
platformdirs = "^2.5.2"
pluggy = "^1.0.0"
poetry-core = "^1.0.8"
ptyprocess = "^0.7.0"
py = "^1.11.0"
pycparser = "^2.21"
PyJWT = "^2.3.0"
pylev = "^1.4.0"
pyparsing = "^3.0.8"
pytest = "^7.1.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.1.0"
python-dateutil = "^2.8.2"
python3-openid = "^3.2.0"
pytz = "^2022.1"
requests = "^2.27.1"
requests-oauthlib = "^1.3.1"
requests-toolbelt = "^0.9.1"
SecretStorage = "^3.3.2"
shellingham = "^1.4.0"
six = "^1.16.0"
social-auth-core = "^4.2.0"
sqlparse = "^0.4.2"
tomli = "^2.0.1"
tomlkit = "^0.10.2"
uritemplate = "^4.1.1"
urllib3 = "^1.26.9"
virtualenv = "^20.14.1"
webencodings = "^0.5.1"
whitenoise = "^6.0.0"
zipp = "^3.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
